Move ssh templates to file_operations module
