+++ app/src/main/java/com/amaze/filemanager/filesystem/cloud/CloudUtil.java
@@ -42,7 +42,7 @@ import com.amaze.filemanager.file_operations.filesystem.filetypes.cloud.onedrive
 import com.amaze.filemanager.filesystem.HybridFile;
 import com.amaze.filemanager.filesystem.HybridFileParcelable;
 import com.amaze.filemanager.filesystem.files.CryptUtil;
-import com.amaze.filemanager.filesystem.ssh.SFtpClientTemplate;
+import com.amaze.filemanager.file_operations.filesystem.filetypes.sftp.templates.SFtpClientTemplate;
 import com.amaze.filemanager.filesystem.ssh.SshClientUtils;
 import com.amaze.filemanager.ui.activities.MainActivity;
 import com.amaze.filemanager.ui.icons.MimeTypes;
